Segregate cli into subcommands
1. Frontend subcommand to start web and rest
2. Rest subcommand to start only rest server
